Move utils/ and change utils.py to utils/package_utils.py
