+++ src/bindings/python/src/openvino/runtime/utils/data_helpers/wrappers/__init__.py
@@ -2,7 +2,9 @@
 # Copyright (C) 2018-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from openvino.utils.data_helpers.data_dispatcher import _data_dispatch
+from openvino._pyopenvino import Tensor, ConstOutput
+from openvino._pyopenvino import InferRequest as InferRequestBase
+
 from openvino.utils.data_helpers.wrappers import tensor_from_file
 from openvino.utils.data_helpers.wrappers import _InferRequestWrapper
 from openvino.utils.data_helpers.wrappers import OVDict
