+++ src/bindings/python/src/openvino/runtime/utils/input_validation/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2018-2024 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+from openvino.utils.input_validation import assert_list_of_ints
+from openvino.utils.input_validation import _check_value
+from openvino.utils.input_validation import check_valid_attribute
+from openvino.utils.input_validation import check_valid_attributes
+from openvino.utils.input_validation import is_positive_value
+from openvino.utils.input_validation import is_non_negative_value
