+++ src/bindings/python/src/openvino/runtime/utils/node_factory/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2018-2024 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+from openvino._pyopenvino import NodeFactory as _NodeFactory
+
+from openvino.runtime import Node, Output, Extension
+from openvino.utils.node_factory import NodeFactory
