+++ src/bindings/python/src/openvino/runtime/utils/reduction/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2018-2024 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+from openvino.runtime import Node
+from openvino.utils.reduction import get_reduction_axes
