+++ src/bindings/python/src/openvino/utils/broadcasting.py
@@ -6,7 +6,7 @@ import logging
 from typing import List, Optional
 
 from openvino.runtime import AxisSet, Node
-from openvino.runtime.utils.types import (
+from openvino.utils.types import (
     NodeInput,
     TensorShape,
     get_dtype,
