+++ src/bindings/python/src/openvino/utils/data_helpers/data_dispatcher.py
@@ -8,7 +8,7 @@ from typing import Any, Dict, Union, Optional
 import numpy as np
 
 from openvino._pyopenvino import ConstOutput, Tensor, Type, RemoteTensor
-from openvino.runtime.utils.data_helpers.wrappers import _InferRequestWrapper, OVDict
+from openvino.utils.data_helpers.wrappers import _InferRequestWrapper, OVDict
 
 ContainerTypes = Union[dict, list, tuple, OVDict]
 ScalarTypes = Union[np.number, int, float]
