+++ src/bindings/python/src/openvino/utils/data_helpers/wrappers.py
