+++ src/bindings/python/src/openvino/utils/decorators.py
@@ -7,7 +7,7 @@ from inspect import signature
 from typing import Any, Callable, Dict, Optional, Union, get_origin, get_args
 
 from openvino.runtime import Node, Output
-from openvino.runtime.utils.types import NodeInput, as_node, as_nodes
+from openvino.utils.types import NodeInput, as_node, as_nodes
 
 
 def _get_name(**kwargs: Any) -> Node:
