+++ src/bindings/python/src/openvino/utils/package_utils.py
