+++ src/bindings/python/src/openvino/utils/reduction.py
